try authorize

+++ src/main/java/com/example/bccom/repo/FlatRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface FlatRepository extends JpaRepository<Flat, Long> {
+public interface FlatRepository extends JpaRepository<Flat, Integer> {
     List<Flat> findByBuilding(Building building);
 }
